pyproject.toml
[tool.poetry]
name = "resume-automation"
version = "0.1.0"
description = "AI-powered resume automation system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false  # Since this is an application, not a library

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"
python-multipart = "^0.0.9"
asyncpg = "^0.29.0"
redis = "^5.0.8"
celery = "^5.4.0"
httpx = "^0.27.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
langchain-community = "^0.3.0"
chromadb = "^0.5.5"
pypdf = "^4.3.1"
python-docx = "^1.1.2"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
aiokafka = "^0.12.0"
psycopg2-binary = "^2.9.10"
email-validator = "^2.2.0"
aiosmtplib = "^4.0.1"
aioboto3 = "^13.2.0"
boto3 = "^1.35.0"
botocore = "^1.35.0"
azure-storage-blob = "^12.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.2.1"
black = "^25.1.0"
ruff = "^0.11.13"
mypy = "^1.16.0"
ipython = "^9.3.0"
rich = "^14.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
